chore: add a `justfile` for automating test workflows
This commit adds a `justfile` for the [Just] command runner. This
performs most of what was previously automated in a big pile of shell
scripts --- hopefully it's a bit nicer to have all this stuff in one
place.

A potential downside of this change is that it does introduce an
additional dependency on `just` for users who want to use the Just
recipes. We could consider alternative approaches like trying to
automate all of these workflows with `cargo` subcommands or something...

[Just]: https://just.systems